added alerts and confirms

+++ src/js/main.js
@@ -1,9 +1,9 @@
 "use strict";
 
-document.addEventListener('DOMContentLoaded', function() {
+document.addEventListener('DOMContentLoaded', function () {
     const logo = document.getElementById("logo");
 
-    logo.addEventListener('click', function() {
+    logo.addEventListener('click', function () {
         window.location.href = "index.html";
     });
 });
@@ -32,7 +32,7 @@ async function displayData() {
         data.forEach(item => { // loopar igenom varje objekt och skriver ut
             const workExperience = document.createElement("div");
             workExperience.classList.add("workexperience");
-            
+
             // målar ut
             workExperience.innerHTML = `
             <h2>Arbetsplats: ${item.companyname}</h2>
@@ -47,9 +47,17 @@ async function displayData() {
             workExperience.appendChild(deleteBtn);
             resultDiv.appendChild(workExperience);
 
-            deleteBtn.addEventListener('click', () => {
-                resultDiv.removeChild(workExperience); // tar bort direkt från sidan
-                deleteData(item.id); // skickar vidare id för att ta bort specifikt
+            deleteBtn.addEventListener('click', async (e) => {
+                try {
+                    const confirmation = confirm("OBS: Är du säker på att du vill radera denna?"); // "är du säker"
+
+                    if (confirmation) {
+                        await deleteData(item.id); // skickar vidare id och inväntar för att ta bort specifikt
+                        resultDiv.removeChild(workExperience); // tar bort resultat direkt från sidan
+                    }
+                } catch (error) {
+                    console.error("Error when deleting data", error);
+                }
             });
         });
     } catch (error) {
@@ -57,24 +65,17 @@ async function displayData() {
     }
 };
 
-function deleteData(id) {
+async function deleteData(id) {
     const url = "https://jeja2306-dt207g-moment2-1.onrender.com/cv/workexperience/" + id; // lägger till angivet id
 
-    fetch(url, {
+    const response = await fetch(url, {
         method: 'DELETE' // metod delete för att radera
-    })
-        .then(response => {
-            if (!response.ok) {
-                throw new Error("Failed to delete data");
-            }
-            return response.json();
-        })
-        .then(data => {
-            console.log("Data deleted", data);
-        })
-        .catch(error => {
-            console.error("Error when deleting data", error);
-        });
+    });
+    if (!response.ok) {
+        throw new Error("Failed to delete data");
+    }
+    const data = await response.json();
+    console.log("Data deleted", data);
 };
 
 displayData(); // hämtar och visar vid start av sida
